+++ lib/braintree/payment_instrument_type.rb
@@ -8,7 +8,6 @@ module Braintree
     VenmoAccount = 'venmo_account'
     UsBankAccount = 'us_bank_account'
     VisaCheckoutCard = 'visa_checkout_card'
-    MasterpassCard = 'masterpass_card' # Deprecated
     SamsungPayCard = 'samsung_pay_card'
     LocalPayment = 'local_payment'
     PayPalHere = 'paypal_here'