+++ spec/integration/braintree/customer_spec.rb
@@ -974,21 +974,6 @@ describe Braintree::Customer do
       android_pay_card.is_network_tokenized?.should == true
     end
 
-    it "returns associated amex express checkout cards" do
-      result = Braintree::Customer.create(
-        :payment_method_nonce => Braintree::Test::Nonce::AmexExpressCheckout
-      )
-      result.success?.should == true
-
-      found_customer = Braintree::Customer.find(result.customer.id)
-      found_customer.amex_express_checkout_cards.size.should == 1
-      found_customer.payment_methods.size.should == 1
-      amex_express_checkout_card = found_customer.amex_express_checkout_cards.first
-      amex_express_checkout_card.should be_a Braintree::AmexExpressCheckoutCard
-      amex_express_checkout_card.token.should_not be_nil
-      amex_express_checkout_card.expiration_year.should_not be_nil
-    end
-
     it "returns associated venmo accounts" do
       result = Braintree::Customer.create(
         :payment_method_nonce => Braintree::Test::Nonce::VenmoAccount