+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Deprecate Masterpass and American Express Checkout payment methods.
 * Add `installment_count` accessor to `dispute.transaction` on dispute webhooks
 * Breaking changes:
-  * Remove support for Transparent Redirect, Coinbase and iDEAL
+  * Remove support for Transparent Redirect, Coinbase, iDEAL, Amex Express Checkout and Masterpass
   * Remove error codes:
     * EuropeBankAccount.IBANIsRequired
     * EuropeBankAccount.BICIsRequired